+++ packages/api/app/distribution-routes.js
@@ -1,12 +1,68 @@
 const router = require('express-promise-router')();
+const { randomId } = require('@cumulus/common/test-utils');
+const { RecordDoesNotExist } = require('@cumulus/errors');
 const {
   handleLoginRequest,
   handleLogoutRequest,
   rootRouter,
 } = require('../endpoints/distribution');
+const { handleCredentialRequest } = require('../endpoints/s3credentials');
 const displayS3CredentialInstructions = require('../endpoints/s3credentials-readme');
-
 const version = require('../endpoints/version');
+const { isAccessTokenExpired } = require('../lib/token');
+const {
+  getConfigurations,
+  isPublicRequest,
+} = require('../lib/distribution');
+
+/**
+ * Ensure request is authorized through EarthdataLogin or redirect to become so.
+ *
+ * @param {Object} req - express request object
+ * @param {Object} res - express response object
+ * @param {Function} next - express middleware callback function
+ * @returns {Promise<Object>} - promise of an express response object
+ */
+async function ensureAuthorizedOrRedirect(req, res, next) {
+  // Skip authentication for debugging purposes.
+  if (process.env.FAKE_AUTH) {
+    req.authorizedMetadata = { userName: randomId('username') };
+    return next();
+  }
+
+  if (isPublicRequest(req.path)) {
+    req.authorizedMetadata = { userName: 'unauthenticated user' };
+    return next();
+  }
+
+  const {
+    accessTokenModel,
+    oauthClient,
+  } = await getConfigurations();
+
+  const redirectURLForAuthorizationCode = oauthClient.getAuthorizationUrl(req.path);
+  const accessToken = req.cookies.accessToken;
+
+  if (!accessToken) return res.redirect(307, redirectURLForAuthorizationCode);
+
+  let accessTokenRecord;
+  try {
+    accessTokenRecord = await accessTokenModel.get({ accessToken });
+  } catch (error) {
+    if (error instanceof RecordDoesNotExist) {
+      return res.redirect(307, redirectURLForAuthorizationCode);
+    }
+
+    throw error;
+  }
+
+  if (isAccessTokenExpired(accessTokenRecord)) {
+    return res.redirect(307, redirectURLForAuthorizationCode);
+  }
+
+  req.authorizedMetadata = { userName: accessTokenRecord.username };
+  return next();
+}
 
 const locate = (req, res) => res.status(501).end();
 
@@ -14,15 +70,13 @@ const profile = (req, res) => res.send('Profile not available.');
 
 const pubkey = (req, res) => res.status(501).end();
 
-const s3Credentials = (req, res) => res.status(501).end();
-
 router.get('/locate', locate);
 router.get('/login', handleLoginRequest);
 router.get('/logout', handleLogoutRequest);
 router.get('/profile', profile);
 router.get('/pubkey', pubkey);
-router.get('/s3Credentials', s3Credentials);
-router.get('/s3CredentialsREADME', displayS3CredentialInstructions);
+router.get('/s3credentials', ensureAuthorizedOrRedirect, handleCredentialRequest);
+router.get('/s3credentialsREADME', displayS3CredentialInstructions);
 // Use router.use to leverage custom version middleware
 router.use('/version', version);