+++ packages/api/lib/distribution.js
@@ -42,6 +42,36 @@ const buildOAuthClient = async () => {
 };
 
 /**
+ * Helper function to pull bucket out of a path string.
+ * Will ignore leading slash.
+ * "/bucket/key" -> "bucket"
+ * "bucket/key" -> "bucket"
+ *
+ * @param {string} path - express request path parameter
+ * @returns {string} the first part of a path which is our bucket name
+ */
+function bucketNameFromPath(path) {
+  return path.split('/').filter((d) => d).shift();
+}
+
+/**
+ * Reads the input path and determines if this is a request for public data
+ * or not.
+ *
+ * @param {string} path - req.path paramater
+ * @returns {boolean} - whether this request goes to a public bucket
+ */
+function isPublicRequest(path) {
+  try {
+    const publicBuckets = process.env.public_buckets.split(',');
+    const requestedBucket = bucketNameFromPath(path);
+    return publicBuckets.includes(requestedBucket);
+  } catch (error) {
+    return false;
+  }
+}
+
+/**
  * Returns a configuration object
  *
  * @returns {Object} the configuration object needed to handle requests
@@ -172,4 +202,6 @@ module.exports = {
   ensureAuthorizedOrRedirect,
   getConfigurations,
   useSecureCookies,
+  bucketNameFromPath,
+  isPublicRequest,
 };